Remove traiiling slashes from sonar.url
In case we have two consecutive slashes in the full_url when querying for the groups in Sonar, the system will respond with a page showing the loading spinner. This will obviously not be acceptable for the JSON parser. To avoid this situation, we should handle a common configuration "error" and remove tailing slashes from the sonar.url.